Extract name fields
