+++ controllers/apply/awards-for-all/fields/index.js
@@ -12,6 +12,7 @@ module.exports = {
     fieldContactLanguagePreference: require('./contact-language-preference'),
     fieldEducationNumber: require('./education-number'),
     fieldMainContactAddress: require('./main-contact-address'),
+    fieldMainContactName: require('./main-contact-name'),
     fieldOrganisationAddress: require('./organisation-address'),
     fieldOrganisationLegalName: require('./organisation-legal-name'),
     fieldOrganisationStartDate: require('./organisation-start-date'),
@@ -26,6 +27,7 @@ module.exports = {
     fieldProjectPostcode: require('./project-postcode'),
     fieldProjectTotalCosts: require('./project-total-costs'),
     fieldSeniorContactAddress: require('./senior-contact-address'),
+    fieldSeniorContactName: require('./senior-contact-name'),
     fieldSeniorContactRole: require('./senior-contact-role'),
     fieldTotalIncomeYear: require('./total-income-year'),
     fieldYourIdeaCommunity: require('./your-idea-community'),
